Factor out Reflection.Emit into an extension
In order to build Reflection.Emit we had to cut some APIs that were
dependent on TypeInfo because it doesn't have a visible constructor.

To workaround for now I changed inheritance to come from Type and
commented out some overrides. See //TYPEINFO in reflection.emit.cs
for the changes made.