Effectively seal IsolatedStorage type in netstandard ref
Since we removed CAS we also removed an abstract member on IsolatedStorage
(GetPermission). Which means if someone derives from IsolatedStorage and
tries to run on a platform, like .NET Framework the would fail because
they wouldn't have implemented that abstract method.

Instead of adding the full closure of IsolatedStoragePermission to
netstandard we are going to make the constructor internal to prevent
anyone from deriving from it. It is believe that the only type to
derive from it is IsolatedStorageFile which lives right next to it
so this seems like the most reasonable compromise.

Fixes https://github.com/dotnet/standard/issues/11