Remove S.R.InteropServices.RuntimeInformation from compatshims
System.Runtime.InteropServices.RuntimeInformation is an OOB library
and isn't part of netstandard. It was referenced by the 1.6 NETStandard.Library
meta-package so we might consider including a reference to it from the 2.0
one but for now we are going to consider it an OOB library that needs
to be referenced independently.