Make netstandard2.0 API compatible with netstandard1.x
This change adds all the APIs that were supported in netstandard1.x
and were part of the NETStandard.Library 1.6 meta-package closure.

By doing this we are introducing some APIs that aren't going to be
compatible with .NET 4.6.1. If they are used they will throw
MissingType or MissingMethod exceptions. We chose to do be back compatible
with netstandard1.x over 100% compatible with .NET 4.6.1. To see the
incompatible APIs refer to ApiCompatBaseline.net461.txt.

To help mitigate people using these APIs we have decided to mark them
with EditorBrowsableState.Never which will hide them in intellisense.
Once the target platforms support these then we will remove those markers
and expose them to intellisense.