Fix Compat shims.
GenFacades (rather CCI) doesn't like it when you give it a set
of assemblies which has conflicting core assemblies. In the shim
generation we were mising mscorlib and system.runtime references
to generate shims for and that caused a number of them to be generated
incorrectly. To fix this we now run genfacades twice once for
mscorlib based stuff and onces for system.runtime based stuff.

Of course now we nave a new core assembly in town with netstandard
which causes other interesting things to show up in the ApiCompat
runs which is why all the baselines for netstandard show failure
to load System.Private.CoreLib (yet another core assembly). The reason
the implementation facades are even looking at System.Private.CoreLib
is because CCI couldn't figure out that netstandard is supposed to
be the core assembly for these facades and so it feel back to
typeof(object).Assembly which is S.P.CoreLib when running on core. This
doesn't really impact the results so just ignoring in the baselines.

This change also factors out logic to get the refs into depproj's insteaad
of manually doing it in the shims.proj. Needed to also turn shim.proj
into a traversal project to make sure it always builds its dependencies
before building.