Use simple name for identifying reference conflicts
References may take the form of a path to a file: full or relative,
simple name, or strong name.

For determining conflicts we're mainly concerned with those that
come from files: NuGet.  As such we'll only examine file references
for conflicts, and we'll make the simplifying assumption that we only
care about simple name.

We could do something much more complicated here that looks
at assembly strong name of all assemblies after resolving them to
files, but in order to do so we'd need to duplicate RAR's logic for
AssemblySearchPaths as well as hit the disk for all files.  If/when
this conflict resolution is promoted to RAR I think it should do that,
but for now we're scoping this down to address the cases that
matter most for packages.