Fix publish for NETCore.App
Publish uses a different target that reads data from the assets file
independently of ResolvePackageDependenciesForBuild.  Hook that
and modify it's output as well.  It happens to use different metadata
for target path, so consider this.

I also found that we were using the NuGet2 sequencing and getting
lucky during build that we'd run early enough.  This was happening
because the SDK targets were being imported *after* our package targets.
This seems wrong so I've filed a few issues, in the meantime I've used a
hacky workaround to condition our targets based on properties we expect
the project to define.