Handle conflicts with external runtime items
In the case of .NETCore shared framework many files exist that should be
considered as runtime files for conflicts.  Copy-local files should only
persist if they win the conflict with these external items.

To avoid doing redundant work I've wrapped items in a type that lazily
instantiates conflict-related data and reuses it.