Rename External to Platform and add manifest
I've renamed external items to platform items since this is more
consistent with other naming.

I've added a platform manifest as a lighter-weight way to specify the
files that make up the shared framework.  With NETCore.App the props
file was over 50KB and 150+ items which is a lot of noise to add to
every portable NETCoreApp project.  This was just quick and dirty,
we could further reduce the size with tokenization/structure.