Trim Deps file
Add a target to read publish-only items from the lock file and
consider these items for conflicts.

Persist the conflicts then remove them from the deps file.

RemoveDepsFileConflicts and _GetLockFileAssemblies are temporary I
expect these to be rewritten in the SDK.  I've tried to factor the
conflict resolution algorithm to be more reusable.