Exclude APIs listed in apiExcludeList.txt when generating Xamarin platform sources
This makes it easier to just pass all the assemblies to GenAPI.exe
without needing to manually filter out the APIs we don't need afterwards.

I also updated seed.cmd to use the in-repo `dotnet` and the platform
assemblies that are installed by the Xamarin installer instead
of some MS internal file share, as well as added a bash version.