Switch netstandard facade project to use a generated type forward list
Fixes https://github.com/dotnet/standard/issues/181

This should fix the fact that GenFacades is invalidating the fully
signed binary.

Instead we now use GenAPI to generate a list of type forwards in
cs form that can just be passed to the compiler and we no longer need
to run the genfacades rewriter.