Fix capitalization issue breaking build on Linux
The file is named GenAPI.exe. On case insensitive file systems it
works if called as GenApi.exe. On linux (and other operating systems)
where capitalization matters, the name mismatch causes the build steps
using this file to fail. This breaks the build.

The build breakage was introduced by e8a2b897.