Update dependency list to account for OOB versions
Previously I omitted version from this list which was causing us to use
any version for inbox decisions, then we'd always raise the dependency
to the baseline version.

This adds version for places where it matters: where we'd have a newer
version than what was inbox and expect the package to bring the out-
of-band dll.