Update file version to be 4.6 which matches the seed assemblies
We need the shim file versions to be higher then the seed assemblies
we are using in order for them to win during conflict resolution. The
seeds were in the 4.6.patch versioning where patch version is using a
date seed that we are still using so it will always be higher.