Follow type forwards when traversing file graph
When examining the file graph we previously only looked at assembly refs
and rooted all of them.

This is problematic for an assembly like netstandard.dll which
references many more assemblies than are typically used.

Now we'll identify assemblies that are purely type-forwards, then follow
those typeForwards to the target assembly.