Add logging to cycle detection
User code may have cycles and we shouldn't stack overflow on those.

We detect the cycle when recursing nodes, or when following type
forwards.  We'll never hit this for corefx assemblies since we check
for cycles at build, but its possible that user code could hit it or
folks mixing and matching assemblies.  I don't want this to be the point
of failure so I merely halt when hitting a cycle and continue execution.