Update shims to include type forwards to some extensions
There are a number of extension libraries that build on top of
netstandard that ship inbox on netfx. For these extensions we wanted
to have the necessary type forwards in the netfx shims to allow folks
to unify types between them when working in a NS library. This change
includes those additional type forwards.