Add a set of extensions that build against netstandard
These extensions build against netstandard library and contain
types that ship in netfx so we want the netfx shims that ship
with netstandard to have type forwards to these assemblies.

These refs were generated based on the API surface we ship in the
individual nuget packages for these libraries.

If we end up needing more we can add them to the extensionlist.txt
and point generateextensions.ps1 script to a directory were they
live and generate it.