Cross-compile GetReferencesNETStandard for desktop
This removes the MetadataReader dependency from our desktop build.

I ported this code from ResolveAssemblyReferences in MSBuild.

It uses the CorMetaDataDispenser API on Windows where these COM types are available.

On mono it dubiously uses ReflectionOnlyLoad, which is actually broken since it doesn't
isolate the AppDomain, but no more broken that RAR is.