Use target path in FileNode for trimming logic
This prevents problems caused by dlls with the same name but different
destination directories, such as satellite assemblies. Previously
these would cause a crash in the trimming logic by causing different
files with the same name to be inserted into a dictionary.

The logic to compute the target path was taken from the SDK:
https://github.com/dotnet/sdk/blob/cd25d6bbfc2aac61cc790c511900427ac64fdaa8/src/Tasks/Common/src/ItemUtilities.cs#L117.