Fix warnings in old project system (.NETFramework projects)
When NETStandard.Library was referenced in a .NETFramework project using the old project system,
it would evaluate the project ignoring conditions on ItemGroups.  Apparently this is done to
display the superset of project configurations (eg both debug and release) in the solution heirarchy.
The warnings were a result of the items never being resolved after design time build finished
resolving references (because of course the items were conditioned out for the target framework).

We can workaround this using choose/when.  Apparently the evaluation of the project is bold
enough to ignore the conditions on an itemgroup, but not bold enough to choose one of many
when (or all) clauses. :)