Breakdown APIs by namespace
We don't want to have all the .NET Standard APIs in a single C# file as GitHub
really doesn't like huge files. That's why the API set is split into multiple
files.

Previously, the contents of the .NET Standard reference assembly was split by
the corresponding assembly in .NET Framework, i.e. it had filenames like
"mscorlib.cs" and "System.Core.cs". This isn't very useful as the assembly
factoring of .NET Framework is really completely irrelevant for the API
surface of .NET Standard.

This splits the APIs by namespace.