Add private fields for structs
This fixes #678. In a nutshell, the compiler needs to know whether a struct
has any fields in order to apply definitive assignment rules. While stripping
all private fields from types is generally OK, we can't do this for structs.
Fortunately, for private fields the compiler doesn't really care what they
are, but what their characteristics are. For example:

1. Does the struct have any fields?

2. Does the struct contain any reference types (to validate generic
   instantiations that have the unmanaged constraint)?

3. Does the struct use the generic parameter in a field declaration (to
   validate cyclic layout problems)?

This adds dummy fields to structs to conform to these rules. These aren't
computed separately but are instead taken from .NET Core.

For more details, see this issue in CoreFX:

https://github.com/dotnet/corefx/issues/6185